requirements.txt
requests==2.31.0
numpy==1.26.4
geopandas==0.14.3
fiona==1.9.6
GDAL==3.4.1
pillow==10.2.0
rasterio==1.3.10
tqdm==4.66.2
roboflow==1.1.30
shapely==2.0.3
ultralytics==8.2.23
yolo==0.3.2
torch==2.3.0
